% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Correct spelling.}
\usage{
correct(word, corrector)
}
\arguments{
\item{word}{A token.}

\item{corrector}{A Java instance of SpellCorrector, given by \code{getcorrector}.}
}
\value{
Corrected spelling
}
\description{
Given an instance of SpellCorrector, return the most probably
corrected spelling of the given word.
}

