% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{getcorrector}
\alias{getcorrector}
\title{Retrieve a Java instance of SpellCorrector.}
\usage{
getcorrector(filepath = "big.txt")
}
\arguments{
\item{filepath}{Path of the corpus.}
}
\value{
a Java instance of SpellCorrector
}
\description{
Retrieve a Java instance of SpellCorrector, with the training file
specified. Language model is trained before the instance is returned.
The spell corrector is adapted from Peter Norvig's demonstration.
}

